requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
django-environ==0.11.2
django-redis==5.4.0
redis==5.0.1
gunicorn==21.2.0
pytest==7.4.3
pytest-django==4.7.0
black==23.9.1
isort==5.12.0
pre-commit==3.5.0
drf-spectacular==0.26.5
dj-database-url==2.1.0
django-db-pool==0.1.0
django-health-check==3.17.0
django-filter==23.3
django-cors-headers==4.3.1
whitenoise==6.6.0
